add & doc: finish the openmp version interface and make the readme documentation more clear
