+++ pytorch_extension/run.sh
@@ -1,1 +1,4 @@
+rm -rf build
+rm -rf dist
+rm -rf pytorch_extension.egg-info
 python setup.py install
