+++ readme.md
@@ -1,34 +1,66 @@
-## Sparse Matrix Pytorch Extension
+# Sparse Matrix Pytorch Extension
 
-### Compile
+## Prerequisites
 
-- description
+Before you can compile the PyTorch extension, you need to install the necessary requirements. Run the following command in your terminal:
 
-    in order to run the pytorch extension, you need to compile the pytorch extension first.
+```bash
+pip install -r requirements.txt
+```
 
-- install the requirements
-    ```bash
-    pip install -r requirements.txt
-    ```
+## Compiling the PyTorch Extension
 
-- compile the pytorch extension
-    ```bash
-    cd pytorch_extension
-    python setup.py install
-    ```
+After installing the prerequisites, navigate to the pytorch_extension directory and run the setup file to compile the PyTorch extension:
 
-### Unit Test 
+```bash
+cd pytorch_extension
+bash run.sh
+```
 
-- description
+## Running the Tests
 
-    in order to test the pytorch extension implementation, we use pytest module to test the pytorch extension.
+This project uses the pytest module for unit testing. To test the PyTorch extension implementation, run the following command:
 
-- run the unit test
-    ```bash
-    pytest ./test
-    ```
+```bash
+pytest ./test
+```
 
-### Notes
+## Benchmark Tool
+
+The main functions for benchmarking are `benchmark_functions_single_test`, `generate_and_benchmark_configurations`, and `show_benchmark_results`.
+
+### benchmark_functions_single_test
+This function benchmarks the performance of four different SPMM implementations on a single test configuration. The implementations are:
+
+- `builtin_sparse_mm`: Uses the built-in `torch.sparse.mm` function.
+- `builtin_sparse_mm_extension`: Uses an extension version of `torch.sparse.mm`.
+- `openmp_sparse_mm`: Uses an OpenMP version of `torch.sparse.mm`.
+- `dense_mm`: Converts the sparse matrices to dense and uses the built-in `torch.mm` function.
+
+The function takes a `SparseMatrixTestConfiguration` object and a number of runs as input. It generates two sparse matrices based on the configuration, times the execution of each SPMM implementation, and prints the results.
+
+### generate_and_benchmark_configurations
+This function generates a list of test configurations and benchmarks them using `benchmark_functions_multiple_test`. The configurations are generated based on the input size and density ranges. The function returns the benchmark results.
+
+### show_benchmark_results
+This function takes a list of benchmark results and prints them in a comparative format. The results are expected to be generated by `generate_and_benchmark_configurations`.
+
+### Usage
+Please note that the benchmarking process can be time-consuming, especially for larger matrices and higher densities. 
+
+```python
+def multiple_benchmark_example():
+    """
+    An example of how to benchmark multiple test configurations.
+    """
+    from src.benchmark import generate_and_benchmark_configurations, show_benchmark_results
+    results = generate_and_benchmark_configurations(size_start=5, size_end=500, size_step=50, density_start=1, density_end=10, density_step=4)
+    show_benchmark_results(results)
+```
+
+## Notes
+
+### pytorch extension include issue
 
 - for those which pytorch built-in function didn't include by torch/extension.h, you need to include the right file like
     ```c++
@@ -39,7 +71,7 @@
     #include <ATen/native/StridedRandomAccessor.h>
     ```
 
-### Reference
+## Reference
 
 - pytorch
     - [pytorch extension](https://pytorch.org/tutorials/advanced/cpp_extension.html)
