+++ src/extension.py
@@ -8,3 +8,10 @@ class ExtensionHandler(object):
         Computes the product of two sparse matrices by using the extension version of ``torch.sparse.mm`` function
         """
         return sparse_mm.sparse_mm(sparse_matrix, sparse_matrix_1)
+    
+    @staticmethod
+    def openmp_sparse_mm(sparse_matrix, sparse_matrix_1):
+        """
+        Computes the product of two sparse matrices by using the openmp parallel version of ``torch.sparse.mm`` function
+        """
+        return sparse_mm.openmp_sparse_mm(sparse_matrix, sparse_matrix_1)
